+++ uui/components/navigation/index.ts
@@ -1,2 +1,3 @@
 export * from './MainMenu';
 export * from './Anchor';
+export { MainMenuLogo, MainMenuCustomElement } from '@epam/uui-components';